Addressed Revi's comments
